+++ text_transform.py
@@ -1,69 +1,15 @@
 import re
-import sys
 from pathlib import Path
-from typing import List, Optional
+from typing import List
 
-from src.file_processor import FileProcessor
-from src.matching import SimpleMatching
+from text_transform_utils import get_filenames_of_name_tags
+
 from src.recipes import split_file_with_separators
 from src.search_configuration import SearchConfiguration
 from src.search_configurations_reader import SearchConfigurationsReader
 from src.text_file import TextFile
 from src.text_file_io import TextFileIO
 from src.text_processor import TextProcessor
-
-
-def get_filenames_of_name_tags(files: List[TextFile]):
-
-    name_tag_pattern = SearchConfiguration(
-        name="name_tag", regex_pattern=r'.*name="(?P<tag>[A-Za-z0-9-]*)".*'
-    )
-
-    tag_matchings_of_filenames = {
-        file.path.stem: FileProcessor(file).search_matchings_of_pattern(
-            name_tag_pattern
-        )
-        for file in files
-    }
-
-    def matching_list_to_value_of_match_group_list(
-        matching_list: List[SimpleMatching], match_group_name: str
-    ) -> List[Optional[str]]:
-        return [
-            matching_to_group_name_value(matching, match_group_name)
-            for matching in matching_list
-        ]
-
-    def matching_to_group_name_value(
-        matching: SimpleMatching, match_group_name: str
-    ) -> Optional[str]:
-        if match := name_tag_pattern.regex.match(matching.text):
-            return match.groupdict().get(match_group_name)
-        return None
-
-    name_tags_of_filenames = {
-        filename: matching_list_to_value_of_match_group_list(
-            matching_list=matchings, match_group_name="tag"
-        )
-        for filename, matchings in tag_matchings_of_filenames.items()
-    }
-
-    name_tags_and_filenames = list(
-        map(reversed, name_tags_of_filenames.items())
-    )
-
-    filenames_of_tag_names = {}
-    tag_set = set()
-    for (tag_names, filename) in name_tags_and_filenames:
-        for tag_name in tag_names:
-            filenames_of_tag_names[tag_name] = filename
-            if tag_name not in tag_set:
-                tag_set.add(tag_name)
-            else:
-                print("Should not have repeated tag names")
-                sys.exit()
-    return filenames_of_tag_names
-
 
 INPUT_TEXT_FILE_PATH = Path(".sandbox/CppCoreGuidelines.md")
 INPUT_SEPARATORS_FILE_PATH = Path(".sandbox/separators.json")
